Fix byte padding issue with seg buffer size
Each row of the texture needs to be a mulitple of the
unpack size, which is typically 4 and cannot be changed
in webgl.  So we load the texture a row at a time
using the first part of the next row as padding.
For the last row we need to copy over the contents
into a new buffer of the correct size.